src/requirements.txt
torch>=2.2.1
transformers
performer-pytorch
tensorboard
axial_positional_embedding
local-attention
fsspec
wrapt
opt_einsum
gast
astunparse
psutil
einops
scipy
torchquad
chaospy
google-cloud-storage
tbparse